requirements.txt
requests==2.31.0
jinja2==3.1.2
pyyaml==6.0.1
jsonschema==4.19.2
pytest==7.4.3
pytest-html==4.1.1
rich==13.7.0
click==8.1.7
